feat(mutators): honest score multipliers, real combo bonus, widen level_score u16
Rework the passive mutator system to be truthful about what it does, add a
genuine combo-scoped bonus that was previously dead code, and widen
level_score so the HUD ring actually reflects progression under multipliers.

Contract changes:
- MutatorDef: rename `combo_score_mult_x100` → `score_mult_x100` (flat
  per-move multiplier, same semantics it always had). Append new field
  `combo_bonus_mult_x100` that multiplies a per-move combo bonus of
  `combo_counter * lines_cleared`, fired only on `lines_cleared > 1`.
- scoring.cairo: rewrite the previously-dead `apply_combo_scoring` helper
  to consume the new multiplier and award the bonus; grid_ops now calls it
  after `process_lines_cleared` inside `execute_move_inline`.
- packing.cairo: widen `level_score` from u8 to u16 in the packed
  `run_data` felt; shift every downstream bit position by +8 (115 of 252
  bits now used). `scoring::update_score` caps at 65535 instead of 255.
  `Game::get_level_score()` and `IGameSystem::get_game_data` tuple now
  return u16.
- level.cairo `calculate_stars_for_level`: fix the inverted semantics so
  `>128 = harder` (matches the doc comment), switch to 5% per unit, and
  make the base more forgiving (3-star at ≤50% moves, 2-star at ≤75%).
- Rewrite all 10 zone passive mutators with distinct themes per zone —
  flat score multiplier for consistency zones (Egypt/Greece/Japan/Inca),
  combo bonus multiplier for skill zones (Norse/Persia/Mayan/Tribal/Inca),
  heavier starting rows on later zones, tighter star thresholds on
  harder zones. Tournament Tiki (settings_id 20) unchanged.

Client changes:
- contractModels.ts: RECS schema renamed + new field.
- runDataPacking.ts: mirror the Cairo bit layout, widen level_score mask
  to 16 bits.
- useMutatorDef: rename JS field, surface `comboBonusMultX100`.
- mutatorConfig.ts: refreshed all 10 passive descriptions with the new
  effect lists (flat vs combo-only multipliers spelled out).

Deployment: bumps `dojo_sepolia.toml` seed from `zkube_2_1` → `zkube_2_2`
so the next migrate creates a fresh world. In-flight sepolia games
discarded (testnet cost — RunData bit layout changed, so old packed
felts would be garbled anyway).

168 contract tests pass (adds a level_score u16 round-trip). Client
typecheck + build clean.

+++ contracts/src/models/game.cairo
@@ -137,7 +137,7 @@ pub impl GameImpl of GameTrait {
 
     /// Get level score
     #[inline(always)]
-    fn get_level_score(self: Game) -> u8 {
+    fn get_level_score(self: Game) -> u16 {
         self.get_run_data().level_score
     }
